requirements.txt
django
djangorestframework
psycopg2
pillow
django-jazzmin
drf-yasg
djangorestframework-simplejwt